Refactor loops to use faster types

+++ src/graphics/font/font.c
@@ -8,8 +8,8 @@ void draw_char(unsigned char c, int x, int y) {
 	/* Get the address of the character in the font bitmap. */
 	BitFont charat = &__font_current[c * BITFONT_CHAR_HEIGHT];
 
-	for (int j = 0; j < BITFONT_CHAR_HEIGHT; ++j) {
-		for (int i = 0; i < BITFONT_CHAR_WIDTH; ++i) {
+	for (uint_fast8_t j = 0; j < BITFONT_CHAR_HEIGHT; ++j) {
+		for (uint_fast8_t i = 0; i < BITFONT_CHAR_WIDTH; ++i) {
 			/* If the bit at the current position is set, draw a pixel,
 			 * the color is driven by Render_Setcolor() */
 			if ((*charat & (1 << (7 - i))) != 0)
